applying SO question and purchases#create logic
DESCRIPTION
I changed the purchases.coffee to use Product.serialize() instead of Purchase.serialize() and Purchase.items, then I created the logic of @getPrice() to submit via POST request the price of the purchase.
I started writing the logic for the purchases#create action, re-wrote the before filter and tested that the POST request works, but I found out that a validation error is triggered as the item.purchase_id is nil and item belongs_to purchase.

TASK TO FINISH
1) Fix the validation errors triggered by the missing item.purchase_id
2) Verify that flash messages are shown
3) Verify that object.errors are shown when saving in _error_messages partial
4) Verify that the redirect after save works
5) Verify that the user is logged in