handling of error messages with AJAX POST request
DESCRIPTION
Creation of an after_filter in the application controller to send in the response.headers the error messages connected to the objects validation or the authentication flow. If the user is not authenticated, the after_filter will set the appropriate message, but this solution needs to be refactored in the future.
Main.coffee includes a new Message class that triggers this alerts, but it will fail if the header is null.

TODO
1) Handle the ajaxError logic if no messages are passed
2) improve the after_filter flash_to_headers
3) continue writing the backend logic
4) refactor the code